Move Database object from VO to DTO.
